IPNC WiFI: ARM: ti814x: Fix ti8148's mmc indices
The sd-card is connected to the second mmc host, and not the first one.
Fix this by modifying and adding the relevant mmc base addresses and dma
channels.

Note: The current implementation for ti814x and ti816x relies on
CONFIG_ARCH_TI81* defines. The current patch continues to do so, but
this should be fixed in the future. Otherwise, omap2_plus will break.